requirements.txt
# Core dependencies for real grocery sustainability scoring
flask
flask-cors
requests
python-dotenv
feedparser
google-generativeai
beautifulsoup4
lxml
rapidfuzz

# Additional dependencies for enhanced functionality
opencv-python
google-genai
inference-sdk
pyobjc
mediapipe
ml-dtypes>=0.5.3
torch>=2.2.0
torchvision
torchaudio
python-doctr[torch]
pillow
numpy
coremltools
flask
flask-cors
flask-socketio
eventletrequests
elevenlabs
